Control test: also provoke lock error in finddomainfd()
svn path=/trunk/; revision=1059

+++ tests/control_test.c
@@ -618,6 +618,12 @@ main(void)
 		}
 	}
 
+	/* already closed fd, should give lock error */
+	if ((finddomainfd(fd, present[i], 0) != -1) || (errno != ENOLCK)) {
+		fputs("Trying to lock an already closed fd must fail\n", stderr);
+		error++;
+	}
+
 	unlink(ctrl_testfile);
 
 	error += test_oneliner();
